+++ src/config/config.php
@@ -18,14 +18,15 @@
 		'scm' => 'git',
 
 		// The SSH/HTTPS adress to your repository
-		'repository' => 'https://github.com/vendor/website.git',
+		// Example: https://github.com/vendor/website.git
+		'repository' => '',
 
 		// The repository credentials : you can leave those empty
 		// if you're using SSH or if your repository is public
 		// In other cases you can leave this empty too, and you will
 		// be prompted for the credentials on deploy
-		'username'   => 'foo',
-		'password'   => 'bar',
+		'username'   => '',
+		'password'   => '',
 
 		// The branch to deploy
 		'branch'     => 'master',